requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic-core==2.33.2
PyMySQL==1.1.2
python-dotenv==1.1.1
RapidFuzz==3.14.1
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
aiohttp==3.12.15
gunicorn==23.0.0